GameScreen cleanup
Move IC and Animation to their own classes.
Update private/public status of members and functions.
