+++ AutoSave.txt
@@ -1,4 +1,4 @@
-2 0 0 0 
+2 0 1 0 
 2 
 0 
 0
