+++ src/Launch.java
@@ -4,14 +4,9 @@ import basicbackbonegame2d.Log;
 import java.io.IOException;
 
 /* TODO:
-    - Replace some direct member accesses with getters/setters
-    - Refactor player; player shouldn't be derived from scene.
+    - Replace direct member accesses with getters/setters.
+    - Refactor player; player probably shouldn't be derived from scene.
     - Player should start at a fixed position, even after reloading a scene or starting a new game.
-    - Rework positioning. Need to be able to track and convert between:
-    --- Window location, i.e. relative to the game window
-    --- Viewport location, i.e. position within the viewport (non-padding subsection of the window, i.e. actual game content)
-    --- Scene location. Viewport might only show a small part of entire scene (ex. if camera follows player).
-        Track location within the overall scene.
     - Text boxes. Maybe handle in GameFrame/Scene since this'll be a type of overlay, always some position in viewport.
     - Move simple scenes to a default template/loader.
     --- Ex) Clock's basic params could be in JSON and a parser could initialize a Scene dynamically.
