+++ src/basicbackbonegame2d/GameScreen.java
@@ -20,20 +20,15 @@ import javax.swing.Timer;
 
 public class GameScreen extends JPanel {
 
-    Timer timer;
-
-    public boolean fromTick = false;
-
     public static final int FRAMES_PER_SEC = 5;
+
+    private Timer timer;
+
+    private boolean fromTick = false;
 
     public static enum CameraType {
         GLOBAL, PLAYER
     }
-
-    double viewportWidth;
-    double viewportHeight;
-    double viewportCenterX;
-    double viewportCenterY;
 
     public CameraType cameraType;
 
@@ -43,6 +38,13 @@ public class GameScreen extends JPanel {
 
     /* currently used cursor type */
     private CursorType activeCursorType;
+
+    // The viewport is the actua lwindow into the game scene.
+    // Its dimensions are given in scene units.
+    private double viewportWidth;
+    private double viewportHeight;
+    private double viewportCenterX;
+    private double viewportCenterY;
 
     public static BufferedImage scaleBufferedImage(BufferedImage input, double scale) {
         int widthOut = (int) (input.getWidth() * scale);
@@ -55,85 +57,6 @@ public class GameScreen extends JPanel {
 
         scaleOp.filter(input, output);
         return output;
-    }
-
-    /*
-     * Inner class for storing image info, including location and the actual
-     * buffered image data.
-     */
-    private class ImageContainer extends Image {
-        // String imgPath;
-        // BufferedImage img;
-        int x, y;
-        // int width, height;
-        double scale;
-        int depth; // What the depth of this image is, used for determining what is drawn on top of
-                   // what else
-        String id; // unique identifier for this image, can be used to check if this is already in
-                   // images list
-
-        ImageContainer(String pathName, int xx, int yy, int dpth, String idStr) {
-            super(pathName);
-
-            x = xx;
-            y = yy;
-            depth = dpth;
-            id = idStr;
-            scale = 1.0;
-        }
-
-        String str() {
-            return id + ", " + imgPath + ", @ (" + x + ", " + y + ", " + depth + ")";
-        }
-
-        boolean equals(ImageContainer other) {
-            boolean eq = (x == other.x) && (y == other.y) && (depth == other.depth) && imgPath.equals(other.imgPath)
-                    && id.equals(other.id);
-            Log.trace("Comparing " + str() + " vs. " + other.str() + ", match: " + eq);
-            return eq;
-        }
-
-        boolean equals(int xx, int yy, int dpth, String imagePath, String idStr) {
-            return (x == xx) && (y == yy) && (depth == dpth) && imgPath.equals(imagePath) && id.equals(idStr);
-        }
-
-        /* Not needed for basic image, but Animation will want this. */
-        void update() {
-
-        }
-
-        BufferedImage getImg() {
-            return img;
-        }
-
-        /*
-         * isAvtive will be used to determine if we need to schedule the timer for
-         * processing any Animations. Regular images are never "active" since they're
-         * static pictures.
-         */
-        boolean isActive() {
-            return false;
-        }
-
-        void setScale(double s) {
-            scale = s;
-        }
-
-        double getScale() {
-            return scale;
-        }
-
-        boolean isAnimated() {
-            return false;
-        }
-
-        int getXOffset() {
-            return 0;
-        }
-
-        int getYOffset() {
-            return 0;
-        }
     }
 
     /* Array of all images that contribute to the current screen. */
@@ -152,123 +75,6 @@ public class GameScreen extends JPanel {
      */
     private List<ImageContainer> refreshImages = new ArrayList<>();
 
-    /*
-     * Class to contain an animation. Animations are stored as large images - if
-     * each frame is w x h pixels (width by height), then the animation image is
-     * size (m x w) x (n x h), where m x n = <number of frames>. Upon creation,
-     * we'll read the image into memory, and for every frame we'll draw a small
-     * region of the image.
-     */
-    private class Animation extends ImageContainer {
-
-        private int numXFrames;
-        private int numYFrames;
-        private int xFrameSize;
-        private int yFrameSize;
-
-        private int numFrames;
-        private int curFrame;
-        private Scene.AnimationType animationType;
-
-        /*
-         * Current state of animation: 0 = inactive/finished, 1 = running. May want to
-         * further distinguish been not started / paused / finished in the future.
-         */
-        private boolean active;
-
-        /*
-         * Constructor needs to know how the large each frame is. We'll compute the
-         * width of the number of frames by dividing the total image by the size of a
-         * frame and with the help of getWidth() or getHeight().
-         */
-        Animation(String pathName, int xx, int yy, int frameSizeX, int frameSizeY, Scene.AnimationType aType, int depth,
-                double s, String id) {
-
-            super(pathName, xx, yy, depth, id);
-
-            xFrameSize = frameSizeX;
-            yFrameSize = frameSizeY;
-
-            width = xFrameSize;
-            height = yFrameSize;
-
-            numXFrames = img.getWidth() / xFrameSize;
-            numYFrames = img.getHeight() / yFrameSize;
-            numFrames = numXFrames * numYFrames;
-
-            curFrame = 0;
-            animationType = aType;
-
-            scale = s;
-
-            active = true;
-
-            Log.trace("Created new animation: " + str());
-        }
-
-        /* Draw the current sceen and update info for the next frame to be drawn. */
-        @Override
-        void update() {
-            if (!fromTick) {
-                return; // only update state if due to tick
-            }
-            if (active) {
-                curFrame = (curFrame + 1) % numFrames;
-                if ((curFrame == 0) && (animationType == Scene.AnimationType.ANIMATED_NO_LOOP)) {
-                    active = false;
-                }
-            }
-        }
-
-        /*
-         * Get the current frame and return it (it's a subimage of the overall image).
-         * All of these calculations are handled using the nominal screen size - the
-         * paint method that called this routine will scale the returned (sub)image as
-         * needed.
-         */
-        @Override
-        BufferedImage getImg() {
-            int yIdx = curFrame / numXFrames;
-            int xIdx = curFrame - yIdx * numXFrames;
-
-            Log.trace("Animation frame " + (curFrame + 1) + "/" + numFrames +
-                    ", x " + xIdx + ", y " + yIdx);
-
-            return scaleBufferedImage(img.getSubimage(xIdx * xFrameSize, yIdx * yFrameSize, xFrameSize, yFrameSize),
-                    scale);
-        }
-
-        /*
-         * If animation is in progress, keep reporting active so that timer keeps
-         * firing.
-         */
-        @Override
-        boolean isActive() {
-            return active;
-        }
-
-        @Override
-        boolean isAnimated() {
-            return true;
-        }
-
-        @Override
-        int getXOffset() {
-            int yIdx = curFrame / numXFrames;
-            int xIdx = curFrame - yIdx * numXFrames;
-
-            return xIdx * xFrameSize;
-        }
-
-        @Override
-        int getYOffset() {
-            int yIdx = curFrame / numXFrames;
-
-            return yIdx * yFrameSize;
-        }
-
-    }
-
     public GameScreen() {
         cameraType = CameraType.PLAYER;
         activeCursorType = CursorType.INVALID; // init to invalid so we'll update at first chance
@@ -281,6 +87,57 @@ public class GameScreen extends JPanel {
 
     public CursorType getActiveCursorType() {
         return activeCursorType;
+    }
+
+    public boolean getFromTick() {
+        return fromTick;
+    }
+
+    public void setFromTick(boolean t) {
+        fromTick = t;
+    }
+
+    // Center location (in scene units) of the viewport / camera.
+    public double getViewportCenterX() {
+        return viewportCenterX;
+    }
+
+    public double getViewportCenterY() {
+        return viewportCenterY;
+    }
+
+    // Scene-based dimensions of the viewport / camera, i.e. how much of the scene
+    // is visible.
+    public void setViewportWidth(double w) {
+        viewportWidth = w;
+    }
+
+    public void setViewportHeight(double h) {
+        viewportHeight = h;
+    }
+
+    public double getViewportWidth() {
+        return viewportWidth;
+    }
+
+    public double getViewportHeight() {
+        return viewportHeight;
+    }
+
+    public double viewportToWindowScale() {
+        return Math.min(
+                (GameFrame.width - GameFrame.xPad * 2.0) / getViewportWidth(),
+                (GameFrame.height - GameFrame.yPad * 2.0) / getViewportHeight());
+    }
+
+    public int windowToSceneX(int x) {
+        double windowCenter = GameFrame.width / 2.0;
+        return (int) (getViewportCenterX() + (x - windowCenter) / viewportToWindowScale());
+    }
+
+    public int windowToSceneY(int y) {
+        double windowCenter = GameFrame.height / 2.0;
+        return (int) (getViewportCenterY() + (y - windowCenter) / viewportToWindowScale());
     }
 
     public void addImgToDrawList(String imgPath, int x, int y, int depth, String id) {
@@ -434,7 +291,7 @@ public class GameScreen extends JPanel {
             ImageContainer ic = images.get(icIdx);
 
             /* Update animations (no-op for static images). */
-            ic.update();
+            ic.update(fromTick);
 
             int srcX = ic.x;
             int srcY = ic.y;
@@ -555,51 +412,5 @@ public class GameScreen extends JPanel {
     public void registerMouseListener(BasicBackboneGame2D.gameMouseListener ml) {
         addMouseListener(ml);
         addMouseMotionListener(ml);
-    }
-
-    // Center location (in scene units) of the viewport / camera.
-    public double getViewportCenterX() {
-        return viewportCenterX;
-    }
-
-    public double getViewportCenterY() {
-        return viewportCenterY;
-    }
-
-    // Scene-based dimensions of the viewport / camera, i.e. how much of the scene
-    // is visible.
-    public void setViewportWidth(double w) {
-        viewportWidth = w;
-    }
-
-    public void setViewportHeight(double h) {
-        viewportHeight = h;
-    }
-
-    public double getViewportWidth() {
-        return viewportWidth;
-    }
-
-    public double getViewportHeight() {
-        return viewportHeight;
-    }
-
-    public double viewportToWindowScale() {
-        return Math.min(
-                (GameFrame.width - GameFrame.xPad * 2.0) / getViewportWidth(),
-                (GameFrame.height - GameFrame.yPad * 2.0) / getViewportHeight());
-    }
-
-    public int windowToSceneX(int x) {
-        double windowCenter = GameFrame.width / 2.0;
-        double scaledbg = viewportToWindowScale();
-        double ddd = getViewportCenterX();
-        double dbg = (getViewportCenterX() + (x - windowCenter) / viewportToWindowScale());
-        return (int) (getViewportCenterX() + (x - windowCenter) / viewportToWindowScale());
-    }
-
-    public int windowToSceneY(int y) {
-        double windowCenter = GameFrame.height / 2.0;
-        return (int) (getViewportCenterY() + (y - windowCenter) / viewportToWindowScale());
     }
 }
